+++ src/components/flat-table/flat-table-checkbox/flat-table-checkbox.style.js
@@ -3,7 +3,7 @@ import StyledCheckbox from "../../../__experimental__/components/checkbox/checkb
 import baseTheme from "../../../style/themes/base";
 
 const StyledFlatTableCheckbox = styled.td`
-  ${({ as, theme }) => css`
+  ${({ as, theme, leftPosition, makeCellSticky }) => css`
     ${as === "td" &&
     `
       background-color: ${theme.colors.white};
@@ -39,6 +39,13 @@ const StyledFlatTableCheckbox = styled.td`
       user-select: none;
       vertical-align: middle;
       white-space: nowrap;
+    `}
+
+    ${makeCellSticky &&
+    css`
+      top: auto;
+      left: ${leftPosition}px;
+      position: sticky;
     `}
   `}